pyproject.toml
[project]
name = "dependancy-llm"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Mizuki", email = "mbaba@kanolab.net" }
]
dependencies = [
    "tqdm>=4.66.1",
    "wandb>=0.16.0",
    "transformers>=4.35.1",
    "pyyaml>=6.0.1",
    "fugashi>=1.3.0",
    "ipadic>=1.0.0",
    "pytest>=7.4.3",
    "torch==1.13.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
